Working version for tests
