+++ radio/src/targets/taranis/CMakeLists.txt
@@ -73,6 +73,7 @@ elseif(PCB STREQUAL X7)
   set(BITMAPS_TARGET 9x_bitmaps)
   set(FONTS_TARGET 9x_fonts_1bit)
   set(LCD_DRIVER lcd_driver_spi.cpp)
+  set(GVAR_SCREEN model_gvars.cpp)
 elseif(PCB STREQUAL XLITE)
   set(PWR_BUTTON "PRESS" CACHE STRING "Pwr button type (PRESS/SWITCH)")
   set(CPU_TYPE STM32F2)
